📢 Mise à jour – v1.4.8 (31/01/2025)
🚀 Nouvelles Fonctionnalités
- Ajout d’un bouton permettant de changer directement la langue d’un article depuis la page d’accueil, affiché à côté du titre de l’article.
- Possibilité d’ajouter une note sur une vidéo : via le menu à trois points, il est désormais possible de cliquer sur « Ajouter une note ». La note s’affiche en bas de l’écran, à gauche du bouton de téléchargement. Un appui long permet de supprimer la note.
- Ajout des nouvelles publications parues pour les assemblées dans la section Nouveautés, avec une prise en compte correcte de la date de publication.
- Les vidéos qui ne sont pas présentes dans le catalogue (ex. vidéos des semaines spéciales) peuvent désormais être ouvertes depuis un document ou depuis la vue « Voir les médias » d’un document.

🧩 Améliorations de l’Expérience
- Lors de l’ajout d’une vidéo dans une autre langue aux favoris, celle-ci est désormais enregistrée dans la bonne langue. Le même comportement est appliqué à l’historique, avec un enregistrement correct dans la base de données.
- Le menu « Télécharger » n’est désormais plus affiché dans le menu déroulant des vidéos et des audios : seul le bouton nuage permet le téléchargement. Une fois le média téléchargé, il est toujours possible de supprimer le fichier comme auparavant.
- Correction de la note persistante affichée en bas de l’écran : les couleurs correspondent désormais correctement au thème clair de l’application.
- Affichage correct des descriptions situées sous les carrousels d’images (galeries) dans certains contenus, notamment dans la leçon 21 du livre « Vivez pour toujours ».
- Les notes et les surlignages associés sont désormais affichés correctement dans la note persistante en bas de page lorsqu’ils existent.
- Pour les alertes (points d’actualité), quel que soit la langue sélectionnée, le lien ouvre désormais toujours la page correcte.
- Lors d’un clic sur une publication depuis une recherche effectuée dans un champ de texte, l’application ouvre désormais correctement le bon passage du document.

⚙️ Optimisation
- Amélioration du chargement des pages WebView lorsque les documents sont affichés en mode image.
- Suppression automatique des fichiers de sous-titres et de leur enregistrement en base de données lors de la suppression d’une vidéo, afin de libérer de l’espace de stockage.

🐞 Corrections & Stabilité
- Correction définitive du problème d’affichage des barres supérieure et inférieure lors du changement de page dans le Texte du jour : celles-ci ne disparaissent plus à l’ouverture d’une fenêtre.
- Correction d’un bug où la suppression d’une note persistante en bas de l’écran ne la retirait pas visuellement.
- La catégorie « Nouveautés (assemblées régionales) » n’est désormais plus affichée lorsqu’elle ne contient aucune publication.
- Correction d’un problème empêchant l’affichage des suggestions de mots lorsqu’un document était ouvert depuis un menu.
- Correction des marques-pages dans la Bible : lorsqu’un marque-page est remplacé, il est désormais mis à jour sur la bonne page.
- Correction d’un problème où les sous-titres des vidéos téléchargées ne fonctionnaient plus.
- Correction d’un bug empêchant l’affichage des versets lors de l’utilisation de l’option « Autres versions ».
- Correction d’un problème dans la page d’aperçu de la Bible : il est désormais possible d’ouvrir correctement les versets lorsqu’un seul verset est présent.
- Correction de certains menus déroulants qui s’affichaient derrière la barre de navigation.

📦 Mises à Jour Techniques
- Mise à jour de flutter_local_notifications : 19.5.0 → 20.0.0
- Mise à jour de file_picker : 10.3.8 → 10.3.10
- Mise à jour de dio : 5.9.0 → 5.9.1
- Mise à jour de Flutter : 3.38.5 → 3.38.6

+++ lib/data/databases/media_collections.dart
@@ -41,8 +41,9 @@ class MediaCollections {
     }
 
     List<Map<String, dynamic>> videosResult = await _database.rawQuery('''
-      SELECT Video.*, MediaKey.* FROM Video
+      SELECT Video.*, MediaKey.*, Subtitle.SubtitleId, Subtitle.ModifiedDateTime AS SubtitleModifiedDateTime, Subtitle.Checksum AS SubtitleChecksum, Subtitle.FilePath AS SubtitleFilePath FROM Video
       LEFT JOIN MediaKey ON Video.MediaKeyId = MediaKey.MediaKeyId
+      LEFT JOIN Subtitle ON Video.VideoId = Subtitle.VideoId
       ''');
 
     if (videosResult.isNotEmpty) {
@@ -281,7 +282,7 @@ class MediaCollections {
         "FrameHeight": video.frameHeight,
         "FrameWidth": video.frameWidth,
         "Label": video.label,
-        "SpecialtyDescription": video.specialtyDescription, // à compléter si nécessaire
+        "SpecialtyDescription": video.specialtyDescription,
         "FilePath": video.filePath,
         "Source": video.source,
         "ModifiedDateTime": video.lastModified,
@@ -334,18 +335,32 @@ class MediaCollections {
         where: "MediaKeyId = ?",
         whereArgs: [mediaKeyId],
       );
-    } else if (media is Video) {
-      await _database.delete(
+    } 
+    else if (media is Video) {
+      // Récupérer les id de la vidéo
+      final videoId = await _database.query(
         "Video",
+        columns: ["VideoId"],
         where: "MediaKeyId = ?",
         whereArgs: [mediaKeyId],
       );
 
-      await _database.delete(
-        "Subtitle",
-        where: "VideoId = ?",
-        whereArgs: [mediaKeyId],
-      );
+      if (videoId.isNotEmpty) {
+        final videoIdResult = videoId.first["VideoId"] as int;
+
+        // Supprimer les dépendances
+        await _database.delete(
+          "Video",
+          where: "VideoId = ?",
+          whereArgs: [videoIdResult],
+        );
+
+        await _database.delete(
+          "Subtitle",
+          where: "VideoId = ?",
+          whereArgs: [videoIdResult],
+        );
+      }
     }
 
     // Enfin, supprimer la clé
